+++ libs/react-components/src/community/components/form/hidden-field/examples/hidden-multiselect.tsx
@@ -1,7 +1,7 @@
 import { StoryFn } from '@storybook/react';
 import React from 'react';
 
-import { Col, Row } from '../../../../../tedi/components/grid';
+import { Col, Row } from '../../../../../tedi/components/layout/grid';
 import { Tag } from '../../../tag/tag';
 import { HiddenField } from '../hidden-field';
 
@@ -14,7 +14,7 @@ const keywords = [
   { label: 'Variant F', value: 'F' },
 ];
 
-export const HiddenMultiselect: StoryFn<typeof HiddenField> = (args) => {
+export const HiddenMultiselect: StoryFn<typeof HiddenField> = (_args) => {
   const [state, setState] = React.useState([keywords[0], keywords[1], keywords[2]]);
 
   return (