+++ libs/react-components/src/community/components/table-of-contents/table-of-contents-modal.tsx
@@ -1,7 +1,7 @@
 import cn from 'classnames';
 import React, { useContext } from 'react';
 
-import { Col, Row } from '../../../tedi/components/grid';
+import { Col, Row } from '../../../tedi/components/layout/grid';
 import { useLabels } from '../../../tedi/providers/label-provider';
 import Button from '../button/button';
 import ButtonContent from '../button-content/button-content';